check1

+++ trainer.py
@@ -485,7 +485,7 @@ class BaseTrainer:
 
             if self.convert_to_db_desc:
                 _, ind = gpu_index_flat_for_image_desc.search(
-                    image_descriptor.reshape(1, -1), 1
+                    image_descriptor.reshape(1, -1).astype(np.float32), 1
                 )
                 # image_descriptor = self.all_image_desc_for_db_conversion[int(ind)]
                 image_descriptor = np.mean(
